fix some bugs in logi

+++ htdocs/codehacking/app/Http/Controllers/AdminMediasController.php
@@ -62,16 +62,41 @@ class AdminMediasController extends Controller
     public function deleteMedia(Request $request){
 
 
-        $photos=Photo::findOrFail($request->checkBoxArray);
-        foreach ($photos as $photo){
+        if (isset($request->delete_single)){
 
-            $photo->delete();
 
+            $this->destroy($request->photo);
+            return redirect()->back();
+        }
+
+
+        if (isset($request->delete_all) && !empty($request->checkBoxArray)){
+
+
+
+            $photos=Photo::findOrFail($request->checkBoxArray);
+
+
+
+
+            foreach ($photos as $photo){
+
+                $photo->delete();
+
+            }
+
+
+
+            return redirect()->back();
+
+        }
+        else{
+
+            return redirect()->back();
         }
 
 
 
-        return redirect()->back();
 
 
 //        return 'work';
